test: split test_server.py by HTTP route into 5 files
Splits the 1242-line test_server.py into one file per HTTP route so each
file stays focused and is easier to navigate. Behavior is unchanged --
classes were moved verbatim with imports adjusted; pytest count remains
exactly 230 before and after.

- test_server_health.py    -- /api/health + static/routing
- test_server_manifest.py  -- /api/manifest streaming + supporting
                              infra (classify, resolve, disconnect,
                              stream, watchdog, git-working-tree gate)
- test_server_signature.py -- /api/manifest/signature
- test_server_file.py      -- /api/file?path=
- test_server_cache.py     -- DELETE /api/manifest/cache
